Moved exception tables from asm to rust
Looking at the objdumps I honestly have no idea what's going on here... they look 100%
equivalent, only difference appears to be 3 very small consecutive symbols being allocated
in a slightly different order.  I'm concerned that there might be some relocation silently failing
in the short branches, but the disassembly looks perfect too, and the direction of symbol
movement _should_ be making things _better_ in this case.  It's nearly 2AM though, so my
bafflement could be related to that.

Whatever.  This fixes it, for now.